django_auth_system/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.4
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
